Adjust classes
