+++ lib/widgets/picos_chart_column.dart
@@ -23,6 +23,7 @@ import 'package:intl/intl.dart';
 import 'package:picos/models/abstract_database_object.dart';
 import 'package:picos/models/weekly.dart';
 import 'package:picos/screens/home_screen/overview/widgets/graph_section.dart';
+import 'package:picos/widgets/picos_chart_helper.dart';
 import 'package:syncfusion_flutter_charts/charts.dart';
 import '../../../../models/daily.dart';
 import '../themes/global_theme.dart';
@@ -52,8 +53,8 @@ class PicosChartColumn extends StatelessWidget {
 
   List<ChartSampleData> _prepareChartData() {
     Map<String, DateTime> datesWithValues = isWeekly == true
-        ? ChartHelper.getLastSevenWeeksFromToday()
-        : ChartHelper.getLastSevenDaysWithDates();
+        ? PicosChartHelper.getLastSevenWeeksFromToday()
+        : PicosChartHelper.getLastSevenDaysWithDates();
 
     return datesWithValues.entries.map((MapEntry<String, DateTime> entry) {
       String dayShortText = entry.key;
@@ -61,10 +62,11 @@ class PicosChartColumn extends StatelessWidget {
 
       AbstractDatabaseObject? matchingData = isWeekly == true
           ? (dataList as List<Weekly>?)?.firstWhereOrNull(
-              (Weekly weekly) => ChartHelper.isWithinWeek(weekly.date, date),
+              (Weekly weekly) =>
+                  PicosChartHelper.isWithinWeek(weekly.date, date),
             )
           : (dataList as List<Daily>?)?.firstWhereOrNull(
-              (Daily daily) => ChartHelper.isSameDay(daily.date, date),
+              (Daily daily) => PicosChartHelper.isSameDay(daily.date, date),
             );
 
       double? value;
@@ -108,7 +110,7 @@ class PicosChartColumn extends StatelessWidget {
 
     return Container(
       decoration: BoxDecoration(
-        border: Border.all(color: theme.blue!, width: ChartHelper.width),
+        border: Border.all(color: theme.blue!, width: PicosChartHelper.width),
         borderRadius: BorderRadius.circular(10),
       ),
       child: SfCartesianChart(
@@ -124,11 +126,11 @@ class PicosChartColumn extends StatelessWidget {
         ),
         primaryXAxis: CategoryAxis(
           majorGridLines: const MajorGridLines(width: 0),
-          axisLine:  AxisLine(
+          axisLine: AxisLine(
             color: theme.blue,
           ),
           labelStyle: const TextStyle(
-            color: ChartHelper.colorBlack,
+            color: PicosChartHelper.colorBlack,
           ),
         ),
         primaryYAxis: NumericAxis(isVisible: false),
